fix route

+++ app/Http/Controllers/Backend/UserController.php
@@ -5,56 +5,87 @@ namespace App\Http\Controllers\Backend;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
     public function index()
     {
-        $user = User::all();
-        return view('backend.user', compact('user'));
+        if (Auth::check() && Auth::user()->role == 0) {
+            $user = User::all();
+            return view('backend.user.user', compact('user'));
+        } else {
+            return redirect('index');
+            exit();
+        }
     }
 
     public function detail($id)
     {
-        $user = User::findOrFail($id);
-        return view('backend.user_detail', compact('user'));
+        if (Auth::check() && Auth::user()->role == 0) {
+            $user = User::findOrFail($id);
+            return view('backend.user.user_detail', compact('user'));
+        } else {
+            return redirect('index');
+            exit();
+        }
     }
 
     public function edit($id)
     {
-        $user = User::findOrFail($id);
-        return view('backend.user_edit', compact('user'));
+        if (Auth::check() && Auth::user()->role == 0) {
+            $user = User::findOrFail($id);
+            return view('backend.user.user_edit', compact('user'));
+        } else {
+            return redirect('index');
+            exit();
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $user = User::find($id);
-        $user->name = $request->input('name');
-        $user->username = $request->input('username');
-        $user->emaiL = $request->input('email');
-        $user->role = $request->input('role');
-        $user->address = $request->input('address');
-        $user->phone = $request->input('phone');
-        $user->dob = $request->input('dob');
-        $user->save();
-        return redirect()->route('user');
+        if (Auth::check() && Auth::user()->role == 0) {
+            $user = User::find($id);
+            $user->name = $request->input('name');
+            $user->username = $request->input('username');
+            $user->emaiL = $request->input('email');
+            $user->role = $request->input('role');
+            $user->address = $request->input('address');
+            $user->phone = $request->input('phone');
+            $user->dob = $request->input('dob');
+            $user->save();
+            return redirect()->route('user');
+        } else {
+            return redirect('index');
+            exit();
+        }
     }
 
     public function delete($id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
-        return back();
+        if (Auth::check() && Auth::user()->role == 0) {
+            $user = User::findOrFail($id);
+            $user->delete();
+            return back();
+        } else {
+            return redirect('index');
+            exit();
+        }
     }
 
     public function create()
     {
-        return view('backend.user_create');
+        if (Auth::check() && Auth::user()->role == 0) {
+            return view('backend.user.user_create');
+        } else {
+            return redirect('index');
+            exit();
+        }
     }
 
     public function store(Request $request)
     {
-    //    dd($request);
+        //    dd($request);
         // $rules = ([
 
         //     'name' => ['required', 'string', 'min:8', 'max:255', 'regex:/^[a-z0-9._-]{4,16}$/i', 'unique:users'],
@@ -62,22 +93,25 @@ class UserController extends Controller
         //     'password' => ['required', 'string',  'confirmed']
         // ]);
         // $this->validate($request, $rules);
-        
+        if (Auth::check() && Auth::user()->role == 0) {
+            $user = User::create([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'role' => $request->role,
+                'address' => $request->address,
+                'phone' => $request->phone,
+                'dob' => $request->dob
+            ]);
 
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => bcrypt($request->password),
-            'role' => $request->role,
-            'address' => $request->address,
-            'phone' => $request->phone,
-            'dob' => $request->dob
-        ]);
-
-        if ($user) {
-            return redirect()->route('user')->with('msg', 'Thêm người dùng thành công');
-        }else{
-            dd('error');
+            if ($user) {
+                return redirect()->route('user')->with('msg', 'Thêm người dùng thành công');
+            } else {
+                dd('error');
+            }
+        } else {
+            return redirect('index');
+            exit();
         }
     }
 }
